+++ app/proguard-rules.pro
@@ -18,6 +18,7 @@
 
 -keepclasseswithmembers public class androidx.recyclerview.widget.RecyclerView { *; }
 
+-keep class com.beemdevelopment.aegis.ui.fragments.*
 -keep class com.beemdevelopment.aegis.importers.** { *; }
 
 -keep class * extends com.google.protobuf.GeneratedMessageLite { *; }
