+++ app/src/androidTest/java/com/beemdevelopment/aegis/OverallTest.java
@@ -127,6 +127,7 @@ public class OverallTest extends AegisTest {
 
         openContextualActionModeOverflowMenu();
         onView(withText(R.string.action_settings)).perform(click());
+        onView(withId(androidx.preference.R.id.recycler_view)).perform(RecyclerViewActions.actionOnItem(hasDescendant(withText(R.string.pref_security_group_title)), click()));
         onView(withId(androidx.preference.R.id.recycler_view)).perform(RecyclerViewActions.actionOnItem(hasDescendant(withText(R.string.pref_encryption_title)), click()));
         onView(withId(android.R.id.button1)).perform(click());
