+++ app/src/main/java/com/beemdevelopment/aegis/ui/fragments/MainPreferencesFragment.java
@@ -1,0 +1,13 @@
+package com.beemdevelopment.aegis.ui.fragments;
+
+import android.os.Bundle;
+
+import com.beemdevelopment.aegis.R;
+
+public class MainPreferencesFragment extends PreferencesFragment {
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        super.onCreatePreferences(savedInstanceState, rootKey);
+        addPreferencesFromResource(R.xml.preferences);
+    }
+}
