+++ app/src/main/res/values-de-rDE/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Herausgeber</string>
     <string name="settings">Einstellungen</string>
     <string name="pref_appearance_group_title">Erscheinungsbild</string>
-    <string name="pref_general_group_title">Allgemein</string>
     <string name="pref_security_group_title">Sicherheit</string>
     <string name="pref_tools_group_title">Werkzeuge</string>
     <string name="pref_backups_group_title">Sicherungskopien</string>
