+++ app/src/main/res/values-el-rGR/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Εκδότης</string>
     <string name="settings">Προτιμήσεις</string>
     <string name="pref_appearance_group_title">Εμφάνιση</string>
-    <string name="pref_general_group_title">Γενικά</string>
     <string name="pref_security_group_title">Ασφάλεια</string>
     <string name="pref_tools_group_title">Εργαλεία</string>
     <string name="pref_backups_group_title">Αντίγραφα Ασφαλείας</string>
