+++ app/src/main/res/values-fi-rFI/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Myöntäjä</string>
     <string name="settings">Asetukset</string>
     <string name="pref_appearance_group_title">Ulkoasu</string>
-    <string name="pref_general_group_title">Yleiset</string>
     <string name="pref_security_group_title">Turvallisuus</string>
     <string name="pref_tools_group_title">Työkalut</string>
     <string name="pref_backups_group_title">Varmuuskopiot</string>
