+++ app/src/main/res/values-in-rID/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Penerbit</string>
     <string name="settings">Pilihan</string>
     <string name="pref_appearance_group_title">Tampilan</string>
-    <string name="pref_general_group_title">Umum</string>
     <string name="pref_security_group_title">Keamanan</string>
     <string name="pref_tools_group_title">Alat</string>
     <string name="pref_backups_group_title">Cadangan</string>
