+++ app/src/main/res/values-nl-rNL/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Uitgever</string>
     <string name="settings">Voorkeuren</string>
     <string name="pref_appearance_group_title">Uiterlijk</string>
-    <string name="pref_general_group_title">Algemeen</string>
     <string name="pref_security_group_title">Beveiliging</string>
     <string name="pref_tools_group_title">Hulpmiddelen</string>
     <string name="pref_backups_group_title">Back-ups</string>
