+++ app/src/main/res/values-pt-rBR/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Nome do serviço</string>
     <string name="settings">Preferências</string>
     <string name="pref_appearance_group_title">Aparência</string>
-    <string name="pref_general_group_title">Geral</string>
     <string name="pref_security_group_title">Segurança</string>
     <string name="pref_tools_group_title">Ferramentas</string>
     <string name="pref_backups_group_title">Backups</string>
