+++ app/src/main/res/values-ru-rRU/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Эмитент</string>
     <string name="settings">Предпочтения</string>
     <string name="pref_appearance_group_title">Внешний вид</string>
-    <string name="pref_general_group_title">Общее</string>
     <string name="pref_security_group_title">Безопасность</string>
     <string name="pref_tools_group_title">Инструменты</string>
     <string name="pref_backups_group_title">Резервные копии</string>
