+++ app/src/main/res/values-sk-rSK/strings.xml
@@ -12,7 +12,6 @@
     <string name="issuer">Vydavateľ</string>
     <string name="settings">Predvoľby</string>
     <string name="pref_appearance_group_title">Vzhľad</string>
-    <string name="pref_general_group_title">Všeobecné</string>
     <string name="pref_security_group_title">Zabezpečenie</string>
     <string name="pref_tools_group_title">Nástroje</string>
     <string name="pref_backups_group_title">Zálohy</string>
