+++ app/src/main/res/values-tr-rTR/strings.xml
@@ -10,7 +10,6 @@
     <string name="issuer">Sağlayıcı</string>
     <string name="settings">Ayarlar</string>
     <string name="pref_appearance_group_title">Görünüm</string>
-    <string name="pref_general_group_title">Genel</string>
     <string name="pref_security_group_title">Güvenlik</string>
     <string name="pref_tools_group_title">Araçlar</string>
     <string name="pref_backups_group_title">Yedekler</string>
