+++ app/src/main/res/xml/preferences_backups.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:title="@string/pref_backups_group_title">
+    <androidx.preference.SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:key="pref_backups"
+        android:title="@string/pref_backups_title"
+        android:summary="@string/pref_backups_summary"
+        app:iconSpaceReserved="false"/>
+    <Preference
+        android:key="pref_backups_location"
+        android:title="@string/pref_backups_location_title"
+        android:summary="@string/pref_backups_location_summary"
+        app:iconSpaceReserved="false"/>
+    <Preference
+        android:key="pref_backups_trigger"
+        android:title="@string/pref_backups_trigger_title"
+        android:summary="@string/pref_backups_trigger_summary"
+        app:iconSpaceReserved="false"/>
+    <Preference
+        android:key="pref_backups_versions"
+        android:title="@string/pref_backups_versions_title"
+        app:iconSpaceReserved="false"/>
+</PreferenceScreen>
