+++ app/src/main/res/xml/preferences_behavior.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:title="@string/pref_behavior_group_title">
+    <androidx.preference.SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:key="pref_copy_on_tap"
+        android:title="@string/pref_copy_on_tap_title"
+        android:summary="@string/pref_copy_on_tap_summary"
+        app:iconSpaceReserved="false"/>
+    <androidx.preference.SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:key="pref_highlight_entry"
+        android:title="@string/pref_highlight_entry_title"
+        android:summary="@string/pref_highlight_entry_summary"
+        app:iconSpaceReserved="false"/>
+    <androidx.preference.SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:key="pref_panic_trigger"
+        android:title="@string/pref_panic_trigger_title"
+        android:summary="@string/pref_panic_trigger_summary"
+        app:iconSpaceReserved="false"/>
+</PreferenceScreen>
